Move ServerLock to interface and implement a 'noop' lock for container environments.
