+++ src/main/java/decodes/util/Pdt.java
@@ -25,7 +25,7 @@ import java.io.LineNumberReader;
 import decodes.db.Database;
 import ilex.util.EnvExpander;
 import ilex.util.Logger;
-import ilex.util.ServerLock;
+import ilex.util.FileServerLock;
 import lrgs.common.DcpAddress;
 
 /**
